fix: correctly renamed Toast.tsx to index.tsx
